backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25

# WebSockets
websockets>=12.0

# File Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
reportlab>=4.0.9

# Web Scraping
beautifulsoup4>=4.12.3
requests>=2.31.0
lxml>=5.1.0

# AI Integration (LiteLLM for multi-provider support)
litellm>=1.30.0

# Security
cryptography>=42.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1

# Development & Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.25.2
